+++ tests/testthat/helper-integration.R
@@ -150,7 +150,6 @@ one_cmt_kappa_fn <- function() {
   irmc_proposals <- lapply(seq_along(studies), function(si)
     admixr2:::.adirmcProposal(
       rxMod, pars0$struct, pinfo$sigma_names,
-      pinfo$sigma_is_prop, pinfo$sigma_is_lnorm,
       pars0$omega, omega_expansion = 2,
       studies[[si]], z_list[[si]], output_var,
       params_list[[si]], cores = 1L,
@@ -230,7 +229,6 @@ one_cmt_kappa_fn <- function() {
   irmc_proposals <- lapply(seq_along(studies), function(si)
     admixr2:::.adirmcProposal(
       rxMod, pars0$struct, pinfo$sigma_names,
-      pinfo$sigma_is_prop, pinfo$sigma_is_lnorm,
       pars0$omega, omega_expansion = 2,
       studies[[si]], z_list[[si]], output_var,
       params_list[[si]], cores = 1L,